Improved UI and refactored code
