+++ frontend/src/components/ChangePassword.jsx
@@ -4,7 +4,7 @@ import * as Dialog from "@radix-ui/react-dialog"
 
 import axiosInstance from "@/axios"
 
-import Loader from "@/components/Loader"
+import Loader from "@/components/ui/Loader"
 
 function ChangePassword() {
     const [resetStatus, setResetStatus] = useState()
