+++ frontend/src/components/Dashboard/Nav.jsx
@@ -19,8 +19,8 @@ function DashboardSubNav() {
                     className={`${
                         currentPath === link.path
                             ? "border-b-2 border-gray-800 font-medium text-gray-900 hover:border-gray-800"
-                            : "font-medium text-gray-500 hover:text-gray-900"
-                    } rounded-t-sm border-b-2 border-white px-3 pb-1.5 pt-2 hover:border-gray-100 hover:bg-gray-100 sm:pb-2`}>
+                            : "border-white font-medium text-gray-500 hover:text-gray-900"
+                    } rounded-t-sm border-b-2 px-3 pb-1.5 pt-2 hover:border-gray-100 hover:bg-gray-100 sm:pb-2`}>
                     {link.label}
                 </Link>
             ))}
