+++ frontend/src/components/ui/SuspenseLoader.jsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 function SuspenseLoader() {
     return (
         <div className="fixed inset-0 flex items-center justify-center bg-white/80 backdrop-blur-sm">
