+++ frontend/src/index.css
@@ -32,7 +32,7 @@ body {
 
 @layer components {
     .wrapper {
-        @apply flex flex-col px-4 py-4 sm:flex-row sm:px-8 lg:px-16;
+        @apply flex flex-col px-6 py-4 sm:flex-row sm:px-8 lg:px-16;
     }
 
     .input {
