+++ frontend/src/pages/AuthGoogle.jsx
@@ -3,7 +3,7 @@ import { Link, useSearchParams } from "react-router-dom"
 
 import { userExists } from "@/signals/user"
 
-import Loader from "@/components/Loader"
+import Loader from "@/components/ui/Loader"
 
 function AuthGoogle() {
     const [loader, setLoader] = useState(true)
