+++ frontend/src/pages/Dashboard/Settings.jsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 function DashboardSettings() {
     return (
         <div className="wrapper">
